move workflow state in data
