+++ infinitic-workflowManager-engine/src/main/kotlin/io/infinitic/workflowManager/engine/engines/handlers/MsgHandler.kt
@@ -13,7 +13,7 @@ import io.infinitic.workflowManager.common.data.workflowTasks.WorkflowTask
 import io.infinitic.workflowManager.common.data.workflowTasks.WorkflowTaskId
 import io.infinitic.workflowManager.common.data.workflowTasks.WorkflowTaskInput
 import io.infinitic.workflowManager.common.messages.WorkflowTaskDispatched
-import io.infinitic.workflowManager.common.states.WorkflowState
+import io.infinitic.workflowManager.common.data.states.WorkflowState
 import io.infinitic.workflowManager.engine.engines.WorkflowEngine
 
 abstract class MsgHandler(
