+++ infinitic-workflowManager-engine/src/main/kotlin/io/infinitic/workflowManager/engine/engines/handlers/WorkflowTaskCompletedHandler.kt
@@ -24,7 +24,7 @@ import io.infinitic.workflowManager.common.data.workflows.WorkflowMessageIndex
 import io.infinitic.workflowManager.common.messages.ChildWorkflowCompleted
 import io.infinitic.workflowManager.common.messages.WorkflowCompleted
 import io.infinitic.workflowManager.common.messages.WorkflowTaskCompleted
-import io.infinitic.workflowManager.common.states.WorkflowState
+import io.infinitic.workflowManager.common.data.states.WorkflowState
 import io.infinitic.workflowManager.engine.engines.WorkflowEngine
 
 class WorkflowTaskCompletedHandler(
