+++ infinitic-workflowManager-engine/src/main/kotlin/io/infinitic/workflowManager/engine/storages/WorkflowStateStorage.kt
@@ -1,7 +1,7 @@
 package io.infinitic.workflowManager.engine.storages
 
 import io.infinitic.workflowManager.common.data.workflows.WorkflowId
-import io.infinitic.workflowManager.common.states.WorkflowState
+import io.infinitic.workflowManager.common.data.states.WorkflowState
 
 interface WorkflowStateStorage {
     fun createState(workflowId: WorkflowId, state: WorkflowState)
